+++ custom_components/solarcharger/state_machine/solar_charge.py
@@ -975,7 +975,7 @@ class SolarCharge(ScOptionState):
         # If device can set current, make it harder to exit pause state by raising the requirement to exit pause state.
         # Min workable current enter pause percent = 0%
         # Min workable current exit pause percent = 10% (ie. harder to change from paused to charging state)
-        if run_state == RunState.PAUSED:
+        if run_state == RunState.PAUSE:
             #####################################
             # For exiting out of paused state.
             #####################################
@@ -1019,7 +1019,7 @@ class SolarCharge(ScOptionState):
             median_net_allocated_power = net_allocations.median_value
             adjusted_activation_power, _ = self.get_adjusted_activation_power(run_state)
 
-            if run_state == RunState.PAUSED:
+            if run_state == RunState.PAUSE:
                 #####################################
                 # For exiting out of paused state.
                 #####################################
@@ -1119,9 +1119,9 @@ class SolarCharge(ScOptionState):
     ) -> None:
         """Check if to continue state."""
 
-        if context.state == RunState.CHARGING:
+        if context.state == RunState.CHARGE:
             self._set_is_continue_charge_state(context)
-        elif context.state == RunState.PAUSED:
+        elif context.state == RunState.PAUSE:
             self._set_is_continue_pause_state(context)
 
     # ----------------------------------------------------------------------------
@@ -1151,7 +1151,7 @@ class SolarCharge(ScOptionState):
 
         # Only set charge limit when in charging state because it can turn on the charger.
         # Once set, the latest and correct state can be requested without calling _async_update_ha() first.
-        if state == RunState.CHARGING:
+        if state == RunState.CHARGE:
             await self.async_set_charge_limit_if_required(chargeable, self.running_goal)
 
         context = self._get_context(
@@ -1287,7 +1287,7 @@ class SolarCharge(ScOptionState):
             await self.async_action_state()
             next_state = self.machine_state.state
 
-            if next_state == current_state and current_state == RunState.ENDED:
+            if next_state == current_state and current_state == RunState.END:
                 # Completed "Ended" state. No more states to run.
                 break