+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.10.2 2026-08-09
 ### Breaking change
 - Reminder: Please [delete then re-add the integration](https://github.com/flashg1/SolarCharger/wiki/Configuration#how-to-delete-and-re-add-the-solarcharger-integration) if upgrading from v0.9.0 or prior.
+- Renamed run state names to Start, Initialise, Charge, Pause, Tidy-up and End.
 ### Fix
 ### New feature
 ### Improvement